Initial commit - CDK package for Bedrock Access Gateway
This is an example CDK package for my fork of the aws-samples/bedrock-access-gateway that stores the API key in Secrets Manager instead of SSM or an environment variable.  Additionally, it uses Route53 hosted zones and an ALB with TLS enabled.

## Background
You can create either ssm.StringParameter or ssm.StringListParameters in a CDK app.  These are public (not secret) values. Parameters of type SecureString cannot be created directly from a CDK application; if you want to provision secrets automatically, use Secrets Manager Secrets (see the aws-cdk-lib/aws-secretsmanager package).

https://docs.aws.amazon.com/cdk/api/v2/python/aws_cdk.aws_secretsmanager/Secret.html
https://docs.aws.amazon.com/cdk/api/v2/python/aws_cdk.aws_ssm/README.html#creating-new-ssm-parameters-in-your-cdk-app